Application/python/requirements.txt
protobuf==3.20.3
numpy==1.22.4
tensorflow==2.8.0
tensorflow-datasets==4.9.2
scikit-image
Pillow
matplotlib
pandas